# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: AlfaUnipac GV — Sistema de Gestão Acadêmica API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Login with email and password
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginBody"
      responses:
        "200":
          description: Successful login
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LoginResponse"
        "401":
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /auth/me:
    get:
      operationId: getMe
      tags: [auth]
      summary: Get current authenticated user
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Current user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  # ALUNOS
  /alunos:
    get:
      operationId: listAlunos
      tags: [alunos]
      summary: List all students
      security:
        - bearerAuth: []
      parameters:
        - name: search
          in: query
          schema:
            type: string
        - name: curso
          in: query
          schema:
            type: string
        - name: status
          in: query
          schema:
            type: string
        - name: turno
          in: query
          schema:
            type: string
      responses:
        "200":
          description: List of students
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Aluno"
    post:
      operationId: createAluno
      tags: [alunos]
      summary: Create a new student
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateAlunoBody"
      responses:
        "201":
          description: Created student
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Aluno"
        "400":
          description: Bad request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /alunos/{id}:
    get:
      operationId: getAluno
      tags: [alunos]
      summary: Get a student by ID
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Student
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AlunoDetalhado"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    patch:
      operationId: updateAluno
      tags: [alunos]
      summary: Update a student
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateAlunoBody"
      responses:
        "200":
          description: Updated student
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Aluno"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    delete:
      operationId: deleteAluno
      tags: [alunos]
      summary: Deactivate a student
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deactivated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Aluno"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /alunos/import:
    post:
      operationId: importAlunos
      tags: [alunos]
      summary: Bulk import students from JSON array
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ImportAlunosBody"
      responses:
        "200":
          description: Import result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ImportResult"
        "400":
          description: Bad request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  # DISCIPLINAS
...
  /bap:
    get:
      operationId: listBap
      tags: [bap]
      summary: List BAP records
      security:
        - bearerAuth: []
      parameters:
        - name: mes
          in: query
          schema:
            type: integer
        - name: ano
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: BAP records
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/BapDetalhado"
    post:
      operationId: generateBap
      tags: [bap]
      summary: Generate BAP for a month/year
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/GenerateBapBody"
      responses:
        "200":
          description: Generated BAP records
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GenerateBapResponse"
  /bap/{id}:
    patch:
      operationId: updateBapStatus
      tags: [bap]
      summary: Update BAP payment status
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateBapStatusBody"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BapMensal"
  /bap/historico:
    get:
      operationId: getBapHistorico
      tags: [bap]
      summary: Get BAP history (list of month/year periods)
      security:
        - bearerAuth: []
      responses:
        "200":
          description: History list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/BapPeriodo"

  # RELATORIOS
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error
...
    LoginBody:
      type: object
      properties:
        email:
          type: string
        password:
          type: string
      required:
        - email
        - password
    LoginResponse:
      type: object
      properties:
        token:
          type: string
        user:
          $ref: "#/components/schemas/User"
      required:
        - token
        - user
    User:
      type: object
      properties:
        id:
          type: integer
        email:
          type: string
        nome:
          type: string
        role:
          type: string
          enum: [Admin, Coordenador, Secretaria]
      required:
        - id
        - email
        - nome
        - role

    # Aluno
    Aluno:
      type: object
      properties:
        id:
          type: integer
        nomeCompleto:
          type: string
        cpf:
          type: string
        matricula:
          type: string
        curso:
          type: string
        turno:
          type: string
          enum: [Matutino, Vespertino, Noturno]
        status:
          type: string
          enum: [Ativo, Retido, Trancado, Egresso]
        valorMensalidade:
          type: number
        financiador:
          type: string
        createdAt:
          type: string
      required:
        - id
        - nomeCompleto
        - cpf
        - matricula
        - curso
        - turno
        - status
        - valorMensalidade
        - financiador
        - createdAt
    AlunoDetalhado:
      type: object
      properties:
        id:
          type: integer
        nomeCompleto:
          type: string
        cpf:
          type: string
        matricula:
          type: string
        curso:
          type: string
        turno:
          type: string
          enum: [Matutino, Vespertino, Noturno]
        status:
          type: string
          enum: [Ativo, Retido, Trancado, Egresso]
        valorMensalidade:
          type: number
        financiador:
          type: string
        createdAt:
          type: string
        frequencias:
          type: array
          items:
            $ref: "#/components/schemas/FrequenciaResumo"
        documentos:
          type: array
          items:
            $ref: "#/components/schemas/Documento"
        emRetencao:
          type: boolean
      required:
        - id
        - nomeCompleto
        - cpf
        - matricula
        - curso
        - turno
        - status
        - valorMensalidade
        - financiador
        - createdAt
        - frequencias
        - documentos
        - emRetencao
    CreateAlunoBody:
      type: object
      properties:
        nomeCompleto:
          type: string
        cpf:
          type: string
        matricula:
          type: string
        curso:
          type: string
        turno:
          type: string
          enum: [Matutino, Vespertino, Noturno]
        status:
          type: string
          enum: [Ativo, Retido, Trancado, Egresso]
        valorMensalidade:
          type: number
        financiador:
          type: string
      required:
        - nomeCompleto
        - cpf
        - matricula
        - curso
        - turno
    UpdateAlunoBody:
      type: object
      properties:
        nomeCompleto:
          type: string
        cpf:
          type: string
        matricula:
          type: string
        curso:
          type: string
        turno:
          type: string
          enum: [Matutino, Vespertino, Noturno]
        status:
          type: string
          enum: [Ativo, Retido, Trancado, Egresso]
        valorMensalidade:
          type: number
        financiador:
          type: string
    ImportAlunosBody:
      type: object
      properties:
        alunos:
          type: array
          items:
            $ref: "#/components/schemas/CreateAlunoBody"
      required:
        - alunos
    ImportResult:
      type: object
      properties:
        imported:
          type: integer
        errors:
          type: array
          items:
            type: string
      required:
        - imported
        - errors

    # Disciplina
...
    FrequenciaResumo:
      type: object
      properties:
        turmaId:
          type: integer
        disciplinaNome:
          type: string
        periodo:
          type: string
        totalAulas:
          type: integer
        presencas:
          type: integer
        faltas:
          type: integer
        faltasJustificadas:
          type: integer
        percentualFaltas:
          type: number
        emRetencao:
          type: boolean
      required:
        - turmaId
        - disciplinaNome
        - periodo
        - totalAulas
        - presencas
        - faltas
        - faltasJustificadas
        - percentualFaltas
        - emRetencao

    # Retencao
...
    Documento:
      type: object
      properties:
        id:
          type: integer
        alunoId:
          type: integer
        tipo:
          type: string
          enum: [Atestado, Declaracao_Trabalho, Licenca_Maternidade]
        dataEntrega:
          type: string
        dataInicioPeriodo:
          type: string
        dataFimPeriodo:
          type: string
        arquivoUrl:
          type: ["string", "null"]
        status:
          type: string
          enum: [Pendente, Aprovado, Rejeitado]
        observacao:
          type: ["string", "null"]
        createdAt:
          type: string
      required:
        - id
        - alunoId
        - tipo
        - dataEntrega
        - dataInicioPeriodo
        - dataFimPeriodo
        - status
        - createdAt
...
    BapMensal:
      type: object
      properties:
        id:
          type: integer
        mes:
          type: integer
        ano:
          type: integer
        alunoId:
          type: integer
        curso:
          type: string
        valorMensalidade:
          type: number
        statusPagamento:
          type: string
          enum: [Pendente, Confirmado]
      required:
        - id
        - mes
        - ano
        - alunoId
        - curso
        - valorMensalidade
        - statusPagamento
    BapDetalhado:
      type: object
      properties:
        id:
          type: integer
        mes:
          type: integer
        ano:
          type: integer
        alunoId:
          type: integer
        curso:
          type: string
        valorMensalidade:
          type: number
        statusPagamento:
          type: string
          enum: [Pendente, Confirmado]
        alunoNome:
          type: string
      required:
        - id
        - mes
        - ano
        - alunoId
        - curso
        - valorMensalidade
        - statusPagamento
        - alunoNome
    GenerateBapBody:
      type: object
      properties:
        mes:
          type: integer
        ano:
          type: integer
      required:
        - mes
        - ano
    GenerateBapResponse:
      type: object
      properties:
        generated:
          type: integer
        total:
          type: number
        records:
          type: array
          items:
            $ref: "#/components/schemas/BapDetalhado"
      required:
        - generated
        - total
        - records
    UpdateBapStatusBody:
      type: object
      properties:
        statusPagamento:
          type: string
          enum: [Pendente, Confirmado]
      required:
        - statusPagamento
    BapPeriodo:
      type: object
      properties:
        mes:
          type: integer
        ano:
          type: integer
        totalAlunos:
          type: integer
        totalValor:
          type: number
        status:
          type: string
      required:
        - mes
        - ano
        - totalAlunos
        - totalValor
        - status

    # Relatorios
...